SEQUOIADBMAINSTREAM-4787 【BUG】【new】【story】【S3分段上传对象】指定用户创建桶，执行分段上传操作，强制删除用户后未清理该桶中分段上传记录和上传数据
git-svn-id: http://192.168.20.11/sequoiadb/trunk@42661 6121f079-4a18-41d3-8c30-8870a322f2c2

+++ SequoiaDB/engine/tools/sequoias3/src/main/java/com/sequoias3/taskmanager/UploadScanClean.java
@@ -76,8 +76,10 @@ public class UploadScanClean {
                                 }else {
                                     while (partList.hasNext()) {
                                         Part part = new Part(partList.getNext());
-                                        dataDao.deleteObjectDataByLobId(null,
-                                                part.getCsName(), part.getClName(), part.getLobId());
+                                        if (part.getLobId() != null) {
+                                            dataDao.deleteObjectDataByLobId(null,
+                                                    part.getCsName(), part.getClName(), part.getLobId());
+                                        }
                                     }
                                 }
                             }finally {
